requirements.txt
pytube==15.0.0
youtube-transcript-api==0.6.2
moviepy==1.0.3
torch==2.4.0
torchaudio==2.4.0
coqui_tts==0.24.1
yt_dlp=2024.8.6
